requirements.txt
python-telegram-bot==20.7
requests==2.31.0
sqlalchemy==2.0.23
psycopg2==2.9.9